Spack environments can concretize specs together
fixes #9902
fixes #11095

This PR adds a boolean entry in spec.yaml called 'concretize_together'.
If set to True all the specs in the environment are concretized to be
self-consistent, meaning there will be a single configuration for each
package and a single provider for each virtual dependency.

The default value is False to maintain backward compatibility with the
behavior before this PR.