Detected specs may have extra attributes and be validated
Specs that are detected now need to be created using the
Spec.from_detection constructor. This constructor allows
for extra attributes to be attached to the spec.

Each package can now define a callback to validate detected
specs. If validation fails (i.e. the callback raise
an exception) a warning is prompted and the spec is not
written to packages.yaml