Be backward compatible wrt the old packages.yaml format
If a packages.yaml in the old format is found, Spack
will convert it in memory. This makes any configuration
in the old format usable by Spack.

If something is added to a configuration file in the old
format via cli, Spack will error and point the user to
the command that is needed to update it. The message points
out that any update will not be forward-compatible.