spack.repo: inject configuration object into functions called in spack.solver.asp
The class MockPackageMultiRepo has been removed, since
mocking a repository and the packages in it is unnecessary.
Instead, we introduce a MockRepositoryBuilder class, which can
reconstruct an ephemeral repository of mock packages using
Jinja2 templates.

Unit-tests on MockPackageMultiRepo have been removed too, as
they were testing the behavior of classes that have been deleted.